+++ openmixup/models/heads/__init__.py
@@ -4,6 +4,7 @@ from .cls_mixup_head import ClsMixupHead
 from .cls_mlp_head import (EfficientFormerClsHead, MetaFormerClsHead, LeViTClsHead, StackedLinearClsHead,
                            VanillaNetClsHead)
 from .contrastive_head import ContrastiveHead, HCRHead
+from .dino_head import DINOHead
 from .latent_pred_head import LatentPredictHead, LatentClsHead, LatentCrossCorrelationHead, MoCoV3Head
 from .mim_head import A2MIMHead, MAEPretrainHead, MAEFinetuneHead, MAELinprobeHead, SimMIMHead
 from .multi_cls_head import MultiClsHead
@@ -15,8 +16,8 @@ from .tokenizer_head import BEiTHead, CAEHead
 from .vision_transformer_head import VisionTransformerClsHead
 
 __all__ = [
-    'A2MIMHead', 'AdaptiveMask', 'BEiTHead', 'CAEHead', 'ClsHead', 'ClsMixupHead', 'ContrastiveHead', 'HCRHead',
-    'EfficientFormerClsHead', 'MetaFormerClsHead', 'LeViTClsHead', 'StackedLinearClsHead', 'VanillaNetClsHead',
+    'A2MIMHead', 'AdaptiveMask', 'BEiTHead', 'CAEHead', 'ClsHead', 'ClsMixupHead', 'ContrastiveHead', 'DINOHead',
+    'EfficientFormerClsHead', 'HCRHead', 'MetaFormerClsHead', 'LeViTClsHead', 'StackedLinearClsHead', 'VanillaNetClsHead',
     'LatentPredictHead', 'LatentClsHead', 'LatentCrossCorrelationHead',
     'MoCoV3Head', 'MAEPretrainHead', 'MAELinprobeHead', 'MAEFinetuneHead', 'MAELinprobeHead',
     'MultiClsHead', 'MultiPrototypes', 'NormLinearClsHead', 'PixelMixBlock', 'RegHead',