+++ openmixup/models/necks/__init__.py
@@ -2,11 +2,12 @@ from .conv_necks import ConvNeck
 from .hr_necks import HRFuseScales
 from .mim_neck import BEiTNeck, CAENeck, MAEPretrainDecoder, NonLinearMIMNeck, SimMIMNeck
 from .mlp_necks import (AvgPoolNeck, MaskPoolNeck, GeneralizedMeanPooling, LinearNeck,
-                        RelativeLocNeck, ODCNeck, MoCoV2Neck, NonLinearNeck, SwAVNeck, DenseCLNeck)
+                        RelativeLocNeck, ODCNeck, MoCoV2Neck, NonLinearNeck, SwAVNeck,
+                        DenseCLNeck, DINONeck)
 from .transformer_necks import TransformerNeck
 
 __all__ = [
-    'AvgPoolNeck', 'BEiTNeck', 'CAENeck', 'ConvNeck', 'HRFuseScales', 'DenseCLNeck',
+    'AvgPoolNeck', 'BEiTNeck', 'CAENeck', 'ConvNeck', 'HRFuseScales', 'DenseCLNeck', 'DINONeck',
     'GeneralizedMeanPooling', 'LinearNeck', 'MoCoV2Neck', 'NonLinearNeck', 'ODCNeck', 'RelativeLocNeck',
     'MAEPretrainDecoder', 'MaskPoolNeck', 'NonLinearMIMNeck', 'SimMIMNeck', 'SwAVNeck',
     'TransformerNeck',